requirement.txt
openai
easyocr
python-dotenv
faiss-cpu
pandas
fastapi
uvicorn
setuptools
wheel